apps/voice-api/pyproject.toml
[project]
name = "voice-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts==0.2.2; python_version>='3.13'",
    "fastapi[standard]>=0.120.4",
    "numpy>=2.3.4",
    "pydantic-settings>=2.11.0",
    "soxr>=1.0.0",
    "twilio>=9.8.5",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build]
packages = ["src/voice_api"]
